Try to squeeze out a few more bytes
Since we can no longer use `7zSD.sfx` (it used to link against
`msvcrt.dll` which is available everywhere, but newer Visual C versions
only ever link against `vcruntime140.dll` which is not as widely
available by far, see https://github.com/git-for-windows/git/issues/1127
for details), we need to find other ways to reduce the footprint.

Leaving out the debug information cuts the size roughly in half. Given
that nobody ever needed to debug the self-extracting part of Git for
Windows' PortableGit, this seems like an obvious thing we want to do.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>