Add a .gitattributes file
Usually we would want to commit all the C/C++ files using Unix line
endings, as they are essentially Git's standard.

However, we're working with imported source tarballs, where essentially
all the files have DOS line endings (with a couple semi-random
exceptions such as `Asm/arm64/LzmaDecOpt.S`).

In the interest of making it easy to work with tarballs to come, pretend
that those files are not text files, as far as Git's attributes are
concerned.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>